# Add JSONLayout to the Google benchmark tests